+++ main.js
@@ -3,10 +3,13 @@
 const path = require('path');
 const url = require('url');
 
-const MainWindow = require('./src/main/main_window');
+const MainWindowBuilder = require('./src/main/main_window_builder');
 const appEvents = require('./src/main/app_events');
 const utils = require('./src/common/utils');
 
+const {
+    app
+} = require('electron');
 // Global reference to window
 let win = null;
 
@@ -19,7 +22,7 @@ function initApp() {
                 protocol: 'file:'
             });
 
-            win = new MainWindow()
+            win = new MainWindowBuilder()
                 .setIcon(iconPath)
                 .setIndex(htmlUrl)
                 .initWindow(utils.isDevEnv());
@@ -28,4 +31,32 @@ function initApp() {
     appEvents(createWindow);
 }
 
-initApp();
+app.disableHardwareAcceleration(); // fixes drag images freezes
+app.commandLine.appendSwitch('force-color-profile', 'srgb'); // Fixes messed up color rendering
+app.allowRendererProcessReuse = true; // Default from electron 9, explicitly set to remove warning
+
+const gotTheLock = app.requestSingleInstanceLock();
+
+if (!gotTheLock) {
+    app.quit();
+} else {
+    app.on('second-instance', (event, commandLine, workingDirectory) => {
+        // If command line is being executed while gaucho is running, this callback will be called
+        // commandLine Array of commands [ './gaucho', 'caca' ]
+        //         in development
+        //         Command [
+        //   '/home/angrykoala/Git/gaucho/node_modules/electron/dist/electron',
+        //   '.'
+        // ]
+
+
+        // Someone tried to run a second instance, we should focus our window.
+        if (win) {
+            if (win.isMinimized()) win.restore();
+            win.focus();
+        }
+    });
+
+
+    initApp();
+}