+++ src/app/common/task_importer.js
@@ -2,18 +2,23 @@
 
 const fs = require('fs');
 
+const EnvVariableHandler = require('../api/env_variables_handler');
+
 module.exports = {
-    parseToJson(suites, version) {
+    parseToJson(suites, version, globalEnv) {
         const parsedSuites = suites.map((suite) => suite.getData());
         const content = {
             "suites": parsedSuites,
             "version": version
         };
+        if (globalEnv) {
+            content.globalEnv = EnvVariableHandler.filterInvalidEnvVariables(globalEnv);
+        }
         return JSON.stringify(content);
     },
-    export(filename, suites, version) {
+    export(filename, suites, version, globalEnv) {
         return new Promise((resolve, reject) => {
-            const content = this.parseToJson(suites, version);
+            const content = this.parseToJson(suites, version, globalEnv);
             fs.writeFile(filename, content, (err) => {
                 if (err) reject(err);
                 else resolve();