+++ src/app/components/settings/settings_menu.vue
@@ -1,6 +1,6 @@
 <template>
     <section class="section settings-page" @contextmenu.stop="">
-        <a class="back-button" @click="saveSettings">
+        <a class="back-button" @click="close">
             <span class="icon is-small">
                 <i class="fas fa-arrow-left"></i>
             </span>
@@ -9,19 +9,28 @@
         <div class="columns is-mobile is-centered">
             <div class="column is-two-thirds settings-menu">
                 <h1 class="title settings-title">Settings</h1>
-                <h3 class="settings-subtitle">Display</h3>
-                <div class="settings-menu-section">
-                    <checkbox-item v-model="bottomBar" label="Bottom bar"></checkbox-item>
-                    <checkbox-item v-model="showTimer" label="Show timer"></checkbox-item>
-                    <theme-selector v-model="theme"></theme-selector>
-                </div>
+
+                <settings-section title="Display" :collapsable="false">
+                    <div class="settings-menu-section">
+                        <checkbox-item v-model="bottomBar" label="Bottom bar"></checkbox-item>
+                        <checkbox-item v-model="showTimer" label="Show timer"></checkbox-item>
+                        <theme-selector v-model="theme"></theme-selector>
+                    </div>
+                </settings-section>
+
                 <hr>
-                <h3 class="settings-subtitle">Actions</h3>
-                <settings-actions @resetSettings="resetSettings"></settings-actions>
-                <h3 class="settings-subtitle">Shortcuts</h3>
-                <div class="settings-menu-section">
+
+                <settings-section title="Actions" :collapsable="false">
+                    <settings-actions @resetSettings="resetSettings"></settings-actions>
+                </settings-section>
+
+                <settings-section :title="envVariablesTitle" :collapsable="true">
+                    <env-variables-form v-model="envVariables"></env-variables-form>
+                </settings-section>
+
+                <settings-section title="Shortcuts" :collapsable="true">
                     <shortcuts-cheatsheet></shortcuts-cheatsheet>
-                </div>
+                </settings-section>
             </div>
         </div>
     </section>
@@ -36,39 +45,58 @@ const components = {
     "checkbox-item": require('./checkbox_item.vue'),
     "shortcuts-cheatsheet": require('./shortcuts_cheatsheet.vue'),
     "theme-selector": require('./theme_selector.vue'),
-    "settings-actions": require('./settings_actions.vue')
+    "settings-actions": require('./settings_actions.vue'),
+    "settings-section": require('./settings_section.vue'),
+    "env-variables-form": require('../common/env_variables_form.vue')
 };
 
 module.exports = {
-    data() {
-        return {
-            bottomBar: this.$store.state.userConfig.bottomBar,
-            showTimer: this.$store.state.userConfig.showTimer,
-            theme: this.$store.state.userConfig.theme
-        };
-    },
     components: components,
-    watch: {
-        theme() {
-            this.$store.commit("setTheme", this.theme);
+    computed: {
+        envVariables: {
+            get() {
+                return this.$store.state.tasks.globalEnv;
+            },
+            set(newValue) {
+                this.$store.commit("setGlobalEnv", newValue);
+            }
         },
-        bottomBar() {
-            this.$store.commit("setBottomBar", this.bottomBar);
+        theme: {
+            get() {
+                return this.$store.state.userConfig.theme;
+            },
+            set(newTheme) {
+                this.$store.commit("setTheme", newTheme);
+            }
         },
-        showTimer() {
-            this.$store.commit("setShowTimer", this.showTimer);
+        bottomBar: {
+            get() {
+                return this.$store.state.userConfig.bottomBar;
+            },
+            set(newBottomBar) {
+                this.$store.commit("setBottomBar", newBottomBar);
+            }
+        },
+        showTimer: {
+            get() {
+                return this.$store.state.userConfig.showTimer;
+            },
+            set(newTimer) {
+                this.$store.commit("setShowTimer", newTimer);
+            }
+        },
+        envVariablesTitle() {
+            return `Env Variables (${this.envVariables.length - 1})`;
         }
     },
     methods: {
-        saveSettings() {
-            this._close();
-        },
         resetSettings() {
             this.bottomBar = true;
             this.showTimer = true;
             this.theme = "classic";
         },
-        _close() {
+        close() {
+            this.$store.dispatch("saveGlobalEnvVariables");
             this.$store.commit("toggleSettings");
         }
     }