+++ src/app/components/task/task_form.vue
@@ -1,25 +1,25 @@
 <template>
     <div class="task-form-wrapper">
         <div class="container is-fluid">
-            <div class="field">
-                <label class="label">Name*</label>
+            <task-form-section title="Name*" :collapsable="false">
                 <div class="control">
                     <input v-model="title" class="input" type="text" :maxlength="constants.maxTaskNameLength">
                 </div>
-            </div>
-            <div class="field">
-                <label class="label">Command*</label>
+            </task-form-section>
+
+            <task-form-section title="Command*" :collapsable="false">
                 <div class="control">
                     <textarea v-model="command" class="textarea"></textarea>
                 </div>
-            </div>
-            <div class="field">
-                <label class="label">Path</label>
-                <div class="control">
-                    <input v-model="path" class="input" type="text" placeholder="Defaults to home">
-                </div>
-            </div>
-            <env-variables-form v-model="env"></env-variables-form>
+            </task-form-section>
+
+            <task-form-section title="Path" :collapsable="false">
+                <input v-model="path" class="input" type="text" placeholder="Defaults to home">
+            </task-form-section>
+
+            <task-form-section :title="envVariablesTitle" :collapsable="true">
+                <env-variables-form v-model="env"></env-variables-form>
+            </task-form-section>
             <div class="field is-grouped is-grouped-right">
                 <div class="control">
                     <button :disabled="!canSave" class="button is-primary save-button is-grouped-right" @click="saveTask">Save</button>
@@ -36,7 +36,8 @@ const Task = require('../../common/task');
 const constants = require('../../../common/constants');
 
 const components = {
-    "env-variables-form": require('./env_variables_form.vue')
+    "task-form-section": require('./task_form_section.vue'),
+    "env-variables-form": require('../common/env_variables_form.vue')
 };
 
 module.exports = {
@@ -54,6 +55,9 @@ module.exports = {
     computed: {
         canSave() {
             return Boolean(this.title && this.command);
+        },
+        envVariablesTitle() {
+            return `Env Variables (${this.env.length - 1})`;
         }
     },
     watch: {
@@ -95,13 +99,13 @@ module.exports = {
 </script>
 
 <style lang="scss" scoped>
-.save-button{
+.save-button {
     width: 80px;
 }
-.container{
+.container {
     padding-top: 10px;
 }
-.task-form-wrapper{
+.task-form-wrapper {
     border-bottom-style: solid;
     border-bottom-width: 1px;
     padding-bottom: 15px;