+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+0.10.0 / 2021-01-16
+===================
+
+* Gaucho now runs as a single-instance app
+* Overhaul of menu and context menu options
+    * Adds cut-copy-paste items to context menu
+    * Suite-related actions moved to context menu
+    * "Import suite" is now part of settings
+* Global Env variables can be set in settings
+    * These will be exported/imported along with the suites in "Export Tasks"
+* Env variable form validation
+* Settings shortcuts menu is now collapsed by default
+* Minor fix in "Desktopify"
+* Dependencies updated
+
 0.9.2 / 2020-10-25
 ==================